data-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
